избавимся от шаблончиков

+++ src/vbox/safe_array.cpp
@@ -13,7 +13,7 @@ SafeArray::SafeArray() {
 }
 
 SafeArray::SafeArray(VARTYPE vt, size_t size) {
-	handle = api->pfnSafeArrayCreateVector(vt, 0, (ULONG)size);
+	handle = api->pfnSafeArrayCreateVector(vt, 0, size);
 	if (!handle) {
 		throw std::runtime_error(__PRETTY_FUNCTION__);
 	}
